requirements.txt
altgraph==0.17
asgiref==3.2.10
astroid==2.4.2
boto3==1.14.18
botocore==1.17.18
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
dask==2.14.0
distributed==2.14.0
Django==3.0.8
django-cors-headers==3.3.0
django-crontab==0.7.1
django-filter==2.2.0
Django-Integrator==1.1.1.1
django-mathfilters==1.0.0
django-query-signals==0.0.2
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.17.1
factory-boy==2.12.0
Faker==4.1.1
featuretools==0.13.4
future==0.18.2
graphviz==0.13.2
gunicorn==20.0.4
HeapDict==1.0.1
html5lib==1.1
httpie==2.2.0
idna==2.9
importlib-metadata==1.6.0
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
msgpack==1.0.0
num2words==0.5.10
numpy==1.18.4
packaging==20.4
pandas==1.0.3
pefile==2019.4.18
Pillow==7.2.0
psutil==5.7.0
psycopg2==2.8.5
py2exe==0.9.2.2
Pygments==2.6.1
PyInstaller==3.6
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5
PyQt5-sip
pyqt5-tools
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
reportlab==3.5.44
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
sklearn==0.0
sortedcontainers==2.1.0
sqlparse==0.3.1
tblib==1.6.0
text-unidecode==1.3
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.45.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.4
zict==2.0.0
zipp==3.1.0
